requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
python-multipart>=0.0.9
openpyxl>=3.1.2
boto3>=1.34.0
python-dotenv>=1.0.1
transformers>=4.47.0
accelerate>=1.2.0
bitsandbytes>=0.45.0
Pillow>=10.0.0
pypdfium2>=4.30.0
